Move anti-virus analyzer classes from shared-web to main application
