+++ cosmetics-web/config/initializers/storage.rb
@@ -5,5 +5,5 @@ Rails.application.config.document_analyzers = Rails.application.config.active_st
 Rails.application.config.document_analyzers.append AntiVirusAnalyzer
 Rails.application.config.document_analyzers.append ReadDataAnalyzer
 # MasterAnalyzer is the only one that we pass to active_storage
-Rails.application.config.active_storage.analyzers = [Shared::Web::MasterAnalyzer]
+Rails.application.config.active_storage.analyzers = [MasterAnalyzer]
 Rails.application.config.active_storage.queue = :cosmetics
